requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.8.1
